pyproject.toml
[build-system]
requires = [ "setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datatable-mcp"
version = "1.0.0"
description = "DataTable MCP Server for in-memory tabular data manipulation with external source integration"
readme = "README.md"
keywords = [ "mcp", "datatable", "pandas", "llm", "ai", "claude", "model", "context", "protocol", "server"]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "fastmcp>=2.3.3",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.1",
    "httpx>=0.28.1",
    "sqlalchemy>=2.0.0",
    "pyarrow>=12.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "aiohttp>=3.12.15",
    "python-dotenv>=1.1.1",
    "gspread-asyncio>=2.0.0",
    "motor>=3.7.1",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed"
]

[project.scripts]
datatable-mcp = "main:main"

[tool.setuptools]
packages = [ "core", "datatable_tools"]
py-modules = [ "main"]
